Delete write requests stuff
