--- memtable.go
@@ -1,81 +1,0 @@
-/*
- * Copyright 2020 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package badger
-
-import (
-	"sync"
-	"sync/atomic"
-
-	"github.com/outcaste-io/badger/v3/skl"
-)
-
-type request struct {
-	// Input values
-	Skl     *skl.Skiplist
-	Entries []*Entry
-	Wg      sync.WaitGroup
-	Err     error
-	ref     int32
-}
-
-type handoverRequest struct {
-	skl      *skl.Skiplist
-	callback func()
-	err      error
-	wg       sync.WaitGroup
-}
-
-func (req *request) reset() {
-	req.Entries = req.Entries[:0]
-	req.Wg = sync.WaitGroup{}
-	req.Err = nil
-	req.ref = 0
-}
-
-func (req *request) IncrRef() {
-	atomic.AddInt32(&req.ref, 1)
-}
-
-func (req *request) DecrRef() {
-	nRef := atomic.AddInt32(&req.ref, -1)
-	if nRef > 0 {
-		return
-	}
-	req.Entries = nil
-	requestPool.Put(req)
-}
-
-func (req *request) Wait() error {
-	req.Wg.Wait()
-	err := req.Err
-	req.DecrRef() // DecrRef after writing to DB.
-	return err
-}
-
-type requests []*request
-
-func (reqs requests) DecrRef() {
-	for _, req := range reqs {
-		req.DecrRef()
-	}
-}
-
-func (reqs requests) IncrRef() {
-	for _, req := range reqs {
-		req.IncrRef()
-	}
-}
